WIP Wavefront Parser / Floor Grid
Current development of Cherry's Wavefront Parser. I haven't settled with the way I want to move forward with it, so there might still be some large changes made with it.

Currently the main program is set up to only check if data is properly loaded, and I've opted to not upload the PNG assets included with the test objects.

Right now, launching the program will only render a floor grid and make use of a rudimentary flythrough camera.

TODO:
- Better Camera Controls
- Clean up / Make use of WindowHandler
- Clean up / Make use of InputHandler
- Finalize Wavefront Parser